fix fragility toggle
mapControl was declared as a const instead of a var, which prevented switchMode() from working properly.

This fixes that by using var rather than const, and also moves the Controls and initial layer declarations above switchMode() so we don't have to rely on hoisting to use variables declared later.

Fixes #16