change naming of library view modes
Summary
=======

- changes "White Fragility mode" to "Colonial Mode"
- changes "General mode" to "Standard Mode"
- changes "Worker Pacification Centres" to "Mechanics Institutes"

Explanation
===========
The point of two modes is to invite map users to think about the work libraries do as institutions, and how certain types of cultural organisation are normalised and certain types are othered. In particular, why are "Indigenous Knowledge Centres" different from "public libraries", and precisely *how*?

Calling the non-standard view "White Fragility Mode" implies that the individual using it is at fault, not woke enough, fragile. This is not helpful as an invitation to be thoughtful and consider the language and perspectives that are normalised.

Changing it to 'colonial' mode shifts the emphasis to that of an institutional and structural concern, rather than an individual fault.

The change for Mechanics Institutes follows more reading about them. Their history is complicated, a mixture of genuine self-help and mutual aid, and attempts by capitalists to control their workers.