fix action conflict
The add-contributors action and the update-topo action both want to check out the current branch and add a commit at the same time, which causes conflicts. This adds contributors as part of the update-topo action and prevents the contributor update action running if `library_services_information.csv` has changed in the original commit.